React: add fetch util to have consistent fetching logic
